check file 404

+++ app/Http/Controllers/Api/FileDestroy.php
@@ -9,9 +9,7 @@ use Illuminate\Support\Facades\Validator;
 
 class FileDestroy extends Controller
 {
-    /**
-     * Handle the incoming request.
-     */
+
     public function __invoke(Request $request)
     {
         $validator = Validator::make($request->all(), [
@@ -26,7 +24,15 @@ class FileDestroy extends Controller
         }
         $id = $request->input('file_id');
 
-        $file = FileData::findOrFail($id);
+        $file = FileData::find($id);
+
+        if (!$file) {
+            return response()->json([
+                'status' => 404,
+                'error' => 'file not Found',
+            ]);
+        }
+
         $file->delete();
 
         return response()->json([
